+++ tests/snowpark/test_procedure.py
@@ -1,3 +1,17 @@
+# Copyright (c) 2024 Snowflake Inc.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+# http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
 import json
 from pathlib import Path
 from textwrap import dedent
@@ -58,6 +72,7 @@ def test_deploy_procedure(
         dedent(
             """\
             create or replace procedure MockDatabase.MockSchema.procedureName(name string)
+            copy grants
             returns string
             language python
             runtime_version=3.8
@@ -69,6 +84,7 @@ def test_deploy_procedure(
         dedent(
             """\
             create or replace procedure MockDatabase.MockSchema.test()
+            copy grants
             returns string
             language python
             runtime_version=3.10
@@ -124,6 +140,7 @@ def test_deploy_procedure_with_external_access(
         dedent(
             """\
             create or replace procedure MockDatabase.MockSchema.procedureName(name string)
+            copy grants
             returns string
             language python
             runtime_version=3.8