+++ tests_integration/testing_utils/snowpark_utils.py
@@ -1,3 +1,17 @@
+# Copyright (c) 2024 Snowflake Inc.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+# http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
 from __future__ import annotations
 
 import json
@@ -351,3 +365,30 @@ class SnowparkTestSteps:
                 stage_name=stage_name
             )
         ]
+
+    def set_grants_on_selected_object(
+        self, object_type: str, object_name: str, privillege: str, role: str
+    ):
+        self._setup.sql_test_helper.execute_single_sql(
+            f"GRANT {privillege} ON {object_type} {object_name} TO ROLE {role};"
+        )
+
+    def assert_that_object_has_expected_grant(
+        self,
+        object_type: str,
+        object_name: str,
+        expected_privillege: str,
+        expected_role: str,
+    ):
+        result = self._setup.sql_test_helper.execute_single_sql(
+            f"SHOW GRANTS ON {object_type} {object_name};"
+        )
+        assert any(
+            [
+                (
+                    grant.get("privilege") == expected_privillege.upper()
+                    and grant.get("grantee_name") == expected_role.upper()
+                )
+                for grant in result
+            ]
+        )